Update wininet.rs
When you need to add headers to your request , it's necessary use -1 in dwHeadersLength, and DWORD have the type c_long and that are u32 and usize can't be negative.

+++ src/um/wininet.rs
@@ -1748,7 +1748,7 @@ extern "system" {
     pub fn HttpAddRequestHeadersW(
         hRequest: HINTERNET,
         lpszHeaders: LPCWSTR,
-        dwHeadersLength: DWORD,
+        dwHeadersLength: i8,
         dwModifiers: DWORD,
     ) -> BOOL;
     pub fn HttpEndRequestA(
